+++ docs/changelog.md
@@ -2,12 +2,16 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
-## [1.1.0] - unreleased
+## [1.1.0] - released Apr 2, 2021
 
 - Added ability to specify chrom sizes file, or refgenie genome.
-- Add --yaml_config option
+- Add perturbation will create regions proportional to chromosome size
 - Improve performance of perturbations
-- Add --drop-from-file and --add-from-file options
+- Add --dropfile, --addfile, and --shiftfile options
+- Add --add_valid option
+- Add --yaml-config option
+- Improved testing framework
+- Improved logging messages
 
 ## [1.0.0] - released May 20, 2020